week-3/airflow/requirements.txt
pandas 
google-cloud-storage
pyarrow
sqlalchemy
psycopg2-binary